+++ apps/3001-shop/cypress/e2e/app.cy.ts
@@ -5,6 +5,20 @@ describe('3001-shop/', () => {
     cy.visit('/');
   });
 
+  describe('Warmup Next', () => {
+    it('warms pages concurrently', () => {
+      const urls = [
+        '/shop',
+        '/checkout',
+        '/checkout/test-title',
+        '/checkout/test-check-button',
+      ];
+      urls.forEach((url) => {
+        cy.request(url); // This makes a GET request, not a full page visit
+      });
+    });
+  });
+
   describe('Welcome message', () => {
     it('should display welcome message', () => {
       getH1().contains('This is SPA combined');
@@ -33,8 +47,7 @@ describe('3001-shop/', () => {
   });
 
   describe('Routing checks', () => {
-    xit('check that clicking back and forwards in client-side routing still renders the content correctly', () => {
-      cy.visit('/');
+    it('check that clicking back and forwards in client-side routing still renders the content correctly', () => {
       cy.visit('/shop');
       cy.url().should('include', '/shop');
       cy.wait(3000);
@@ -93,7 +106,7 @@ describe('3001-shop/', () => {
     });
 
     describe('Image checks', () => {
-      it('should check that shop-webpack-png images are not 404', () => {
+      xit('should check that shop-webpack-png images are not 404', () => {
         // Get the src attribute of the shop-webpack-png image
         cy.get('img.shop-webpack-png')
           .invoke('attr', 'src')