+++ packages/manifest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/manifest
 
+## 0.1.9
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.9
+- @module-federation/managers@0.1.9
+- @module-federation/dts-plugin@0.1.9
+
 ## 0.1.8
 
 ### Patch Changes