+++ packages/runtime-tools/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [1.0.1-canary.1](https://github.com/module-federation/universe/compare/runtime-1.0.0...runtime-1.0.1-canary.1) (2023-12-06)
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [5ef0150]
+  - @module-federation/runtime@0.1.9
+  - @module-federation/webpack-bundler-runtime@0.1.9
+
 ## 0.1.8
 
 ### Patch Changes