+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.1.9
+
 ## 0.1.8
 
 ## 0.1.7